+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to Schema Forge will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2025-03-25
+
+1. Publish `jsonSchemaToGeminiOldTool`, `jsonSchemaToGeminiOldResponseSchema` that whould be published in 1.0.0.
+2. Support the tool format and responseSchema format of google vertex API, `@google-cloud/vertexai`. Supply the following functions
+   1. `jsonSchemaToGeminiVertexTool`
+   2. `jsonSchemaToGeminiVertexResponseSchema`
+   3. `classToGeminiVertexResponseSchema`
+   4. `classToGeminiVertexTool`
+
 ## [1.0.1] - 2025-03-25
 
 1. Improve README